feature: Final ContractInteraction flow
