+++ src/routes/safe/components/Balances/SendModal/screens/ContractInteraction/ContractABI/index.jsx
@@ -1,0 +1,29 @@
+// @flow
+import React from 'react'
+
+import TextareaField from '~/components/forms/TextareaField'
+import Col from '~/components/layout/Col'
+import Row from '~/components/layout/Row'
+import EtherscanService from '~/logic/contractInteraction/sources/EtherscanService'
+
+const mustBeValidABI = (abi: string) => {
+  try {
+    const parsedABI = EtherscanService.extractUsefulMethods(JSON.parse(abi))
+
+    if (parsedABI.length === 0) {
+      return 'no data'
+    }
+  } catch (e) {
+    return []
+  }
+}
+
+const ContractABI = () => (
+  <Row margin="sm">
+    <Col>
+      <TextareaField name="abi" placeholder="ABI*" text="ABI*" type="text" validate={mustBeValidABI} />
+    </Col>
+  </Row>
+)
+
+export default ContractABI
